Added the turret hex and fixed ship controlling while in build mode
Turret hex uses shootbehavior with minimal changes, perhaps should be made to inherit instead in the future to  automatically inherit any other changes made to shooting. Also, ships could be controlled while build mode was on; this is no longer the case. However, ending build mode with a right click commands the ships to move there which is not intended.