Shop GUI basic functionality completed
The shop works properly and as intended, and now likely only needs very niche quality of life updates and a visual overhaul. In this version, hexes can be selected from the hotbar through keys, the buttons, or the shop list. The hotbar can be manipulated for both quick and recent slots. The shop can be opened and closed through double clicking the hotbar, the shop button, pressing P, or with the shop X. The search bar does not work.