Basic missile hex functionality implemented
It has a very long range and more damage, but shoots slower (when compared to the turret hex). Missile Hex inherits from TurretHex, making it turn on and off in a similar manner (only differences are the few constants that are changed in Awake()). To make this happen, minor refactoring was done in TurretHex to make the class more general and easier to inherit from.